Export a single macro only
