+++ src/internal/mod.rs
@@ -6,10 +6,6 @@
 pub extern crate smallvec;
 
 #[macro_use]
-pub mod codegen;
-#[macro_use]
-pub mod parse;
-#[macro_use]
 pub mod select;
 
 pub mod channel;
